+++ src/repositories/creator/creator.module.ts
@@ -1,9 +1,9 @@
 import { Module } from '@nestjs/common';
 import { MongooseModule, getModelToken } from '@nestjs/mongoose';
-import { CreatorRepository } from './creator.service';
+import { CreatorRepository } from './creator.repository';
 import { ContentCreator, ContentCreatorSchema } from './schemas/creator.schema';
 import { ConfigModule, ConfigService } from '@nestjs/config';
-import { MockCreatorService } from './creator.service.mock';
+import { MockCreatorService } from './creator.repository.mock';
 import { Model } from 'mongoose';
 
 @Module({
