+++ src/repositories/video/video.module.ts
@@ -1,9 +1,9 @@
 import { Module } from '@nestjs/common';
 import { MongooseModule, getModelToken } from '@nestjs/mongoose';
-import { VideoRepository } from './video.service';
+import { VideoRepository } from './video.repository';
 import { Video, VideoDocument, VideoSchema } from './schemas/video.schema';
 import { ConfigModule, ConfigService } from '@nestjs/config';
-import { MockVideoRepository } from './video.service.mock';
+import { MockVideoRepository } from './video.repository.mock';
 import { Model } from 'mongoose';
 
 @Module({
